attempt to update and clean audio flow with volume scaling

+++ CSoundLib/src/state.c
@@ -3,6 +3,6 @@
 // keeps the state of the current audio session
 audio_state* csoundlib_state;
 
-double lib_getCurrentRmsOutput() {
+float lib_getCurrentRmsOutput() {
     return csoundlib_state->current_rms_ouput;
 }
